railsadmin

+++ db/migrate/20241120132745_drop_fishes_table.rb
@@ -1,6 +1,5 @@
 class DropFishesTable < ActiveRecord::Migration[7.0]
   def change
-    drop_table :fishes
+    drop_table :fishes if ActiveRecord::Base.connection.table_exists?(:fishes)
   end
 end
-
